migrate endpoint to search.py
