+++ server/routes/search.py
@@ -15,6 +15,7 @@
 
 from flask import Blueprint, current_app, request
 import flask
+import os
 import services.datacommons as dc
 
 bp = Blueprint('search', __name__)
@@ -72,3 +73,12 @@ def search_dc():
     return flask.render_template('search_dc.html',
                                  query_text=query_text,
                                  results=results)
+
+
+@bp.route('/tools/rich_search')
+def rich_search():
+    if os.environ.get('FLASK_ENV') == 'production':
+        flask.abort(404)
+    return flask.render_template(
+        'tools/rich_search.html',
+        maps_api_key=current_app.config['MAPS_API_KEY'])
