+++ server/templates/tools/rich_search.html
@@ -18,7 +18,7 @@
 
 {% set is_hide_full_footer = true %}
 {% set subpage_title = "Rich Search" %}
-{% set subpage_url = url_for('tools.rich_search') %}
+{% set subpage_url = url_for('search.rich_search') %}
 {% set title = "Rich Search" %}
 
 {% block head %}
